:truck:  Restructure backend files
