--- issuepoker-backend/src/main/java/de/muenchen/issuepoker/entities/dto/VoteMapper.java
@@ -1,11 +1,0 @@
-package de.muenchen.issuepoker.entities.dto;
-
-import de.muenchen.issuepoker.entities.Vote;
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-
-@Mapper
-public interface VoteMapper {
-    @Mapping(target = "id", ignore = true)
-    Vote toEntity(VoteRequestDTO voteRequestDTO, String username);
-}
