--- issuepoker-backend/src/main/java/de/muenchen/issuepoker/entities/dto/VoteRequestDTO.java
@@ -1,8 +1,0 @@
-package de.muenchen.issuepoker.entities.dto;
-
-import de.muenchen.issuepoker.entities.validation.ValidVote;
-import jakarta.validation.constraints.NotNull;
-
-public record VoteRequestDTO(
-        @NotNull @ValidVote int voting) {
-}
