+++ issuepoker-backend/src/main/java/de/muenchen/issuepoker/entities/vote/request/VoteRequestRevealedDTO.java
@@ -1,4 +1,4 @@
-package de.muenchen.issuepoker.entities.dto;
+package de.muenchen.issuepoker.entities.vote.request;
 
 import jakarta.validation.constraints.NotNull;
