+++ issuepoker-backend/src/main/java/de/muenchen/issuepoker/entities/vote/request/VoteRequestVotingDTO.java
@@ -1,0 +1,8 @@
+package de.muenchen.issuepoker.entities.vote.request;
+
+import de.muenchen.issuepoker.entities.vote.validation.ValidVote;
+import jakarta.validation.constraints.NotNull;
+
+public record VoteRequestVotingDTO(
+        @NotNull @ValidVote int voting) {
+}
