+++ issuepoker-backend/src/main/java/de/muenchen/issuepoker/entities/vote/response/VotesDTO.java
@@ -1,4 +1,4 @@
-package de.muenchen.issuepoker.entities.dto;
+package de.muenchen.issuepoker.entities.vote.response;
 
 import java.util.List;
